move the TestModels.TestLateral from thr test_ext.py to a separate test file test_laterals.py
